word-mcp/requirements.txt
fastmcp>=2.13.2
python-docx>=1.1.0
